+++ lib/whatsapp_sdk/api/messages.rb
@@ -111,7 +111,7 @@ module WhatsappSdk
       def send_image(
         sender_id:, recipient_number:, image_id: nil, link: nil, caption: "", message_id: nil
       )
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "image_id or link is required" if !image_id && !link
+        raise WhatsappSdk::Resource::Errors::MissingArgumentError, "image_id or link is required" if !image_id && !link
 
         params = {
           messaging_product: "whatsapp",
@@ -153,7 +153,7 @@ module WhatsappSdk
         ).returns(WhatsappSdk::Api::Response)
       end
       def send_audio(sender_id:, recipient_number:, audio_id: nil, link: nil, message_id: nil)
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "audio_id or link is required" if !audio_id && !link
+        raise WhatsappSdk::Resource::Errors::MissingArgumentError, "audio_id or link is required" if !audio_id && !link
 
         params = {
           messaging_product: "whatsapp",
@@ -195,7 +195,7 @@ module WhatsappSdk
       def send_video(
         sender_id:, recipient_number:, video_id: nil, link: nil, caption: "", message_id: nil
       )
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "video_id or link is required" if !video_id && !link
+        raise WhatsappSdk::Resource::Errors::MissingArgumentError, "video_id or link is required" if !video_id && !link
 
         params = {
           messaging_product: "whatsapp",
@@ -241,7 +241,10 @@ module WhatsappSdk
       def send_document(
         sender_id:, recipient_number:, document_id: nil, link: nil, caption: "", message_id: nil
       )
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "document or link is required" if !document_id && !link
+        if !document_id && !link
+          raise WhatsappSdk::Resource::Errors::MissingArgumentError,
+                "document or link is required"
+        end
 
         params = {
           messaging_product: "whatsapp",
@@ -283,7 +286,7 @@ module WhatsappSdk
         ).returns(WhatsappSdk::Api::Response)
       end
       def send_sticker(sender_id:, recipient_number:, sticker_id: nil, link: nil, message_id: nil)
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "sticker or link is required" if !sticker_id && !link
+        raise WhatsappSdk::Resource::Errors::MissingArgumentError, "sticker or link is required" if !sticker_id && !link
 
         params = {
           messaging_product: "whatsapp",
@@ -371,7 +374,10 @@ module WhatsappSdk
       def send_interactive_message(
         sender_id:, recipient_number:, interactive: nil, interactive_json: nil, message_id: nil
       )
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "interactive or interactive_json is required" if !interactive && !interactive_json
+        if !interactive && !interactive_json
+          raise WhatsappSdk::Resource::Errors::MissingArgumentError,
+                "interactive or interactive_json is required"
+        end
 
         params = {
           messaging_product: "whatsapp",
@@ -446,7 +452,10 @@ module WhatsappSdk
       def send_template(
         sender_id:, recipient_number:, name:, language:, components: nil, components_json: nil
       )
-        raise WhatsappSdk::Resource::Error::MissingArgumentError, "components or components_json is required" if !components && !components_json
+        if !components && !components_json
+          raise WhatsappSdk::Resource::Errors::MissingArgumentError,
+                "components or components_json is required"
+        end
 
         params = {
           messaging_product: "whatsapp",