+++ lib/whatsapp_sdk/resource/interactive_action_section_row.rb
@@ -61,7 +61,7 @@ module WhatsappSdk
         title_length = title.length
         return if title_length <= ACTION_SECTION_TITLE_MAXIMUM
 
-        raise WhatsappSdk::Resource::Error::InvalidInteractiveActionSectionRow,
+        raise WhatsappSdk::Resource::Errors::InvalidInteractiveActionSectionRow,
               "Invalid length #{title_length} for title in section row. "\
               "Maximum length: #{ACTION_SECTION_TITLE_MAXIMUM} characters."
       end
@@ -74,7 +74,7 @@ module WhatsappSdk
         id_length = id.length
         return if id_length <= ACTION_SECTION_ID_MAXIMUM
 
-        raise WhatsappSdk::Resource::Error::InvalidInteractiveActionSectionRow,
+        raise WhatsappSdk::Resource::Errors::InvalidInteractiveActionSectionRow,
               "Invalid length #{id_length} for id in section row. Maximum length: "\
               "#{ACTION_SECTION_ID_MAXIMUM} characters."
       end
@@ -84,7 +84,7 @@ module WhatsappSdk
         description_length = description.length
         return if description_length <= ACTION_SECTION_DESCRIPTION_MAXIMUM
 
-        raise WhatsappSdk::Resource::Error::InvalidInteractiveActionSectionRow,
+        raise WhatsappSdk::Resource::Errors::InvalidInteractiveActionSectionRow,
               "Invalid length #{description_length} for description in section " \
               "row. Maximum length: #{ACTION_SECTION_DESCRIPTION_MAXIMUM} characters."
       end